+++ goldenmaster/tb_enum/implementation/tests/CMakeLists.txt
@@ -1,0 +1,21 @@
+
+project(tb_enum_tests)
+cmake_minimum_required(VERSION 3.20)
+
+set(CMAKE_INCLUDE_CURRENT_DIR ON)
+
+set(CMAKE_AUTOUIC ON)
+set(CMAKE_AUTOMOC ON)
+set(CMAKE_AUTORCC ON)
+find_package(Qt5 REQUIRED COMPONENTS Core Test)
+
+set(CMAKE_CXX_STANDARD 14)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+
+# append local binary directory for conan packages to be found
+set(CMAKE_MODULE_PATH ${CMAKE_BINARY_DIR} ${CMAKE_MODULE_PATH})
+add_executable(test_tb_enum_enuminterface test_enuminterface.cpp)
+
+
+find_package(tb_enum QUIET COMPONENTS tb_enum_api tb_enum_impl )
+target_link_libraries(test_tb_enum_enuminterface  tb_enum_api tb_enum_impl Qt5::Core Qt5::Test)
