+++ goldenmaster/tb_enum/implementation/tests/test_enuminterface.cpp
@@ -15,86 +15,100 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "tst_nestedstruct3interface.h"
-#include "nestedstruct3interface.h"
+#include "test_enuminterface.h"
+#include "implementation/enuminterface.h"
 
 /**
  *  will be called before the first test function is executed.
 */
-void TestTestbed2NestedStruct3Interface::initTestCase(){
+void TestTbEnumEnumInterface::initTestCase(){
 
 }
 
 /**
  *  will be called to create a global test data table.
 */
-void TestTestbed2NestedStruct3Interface::initTestCase_data(){
+void TestTbEnumEnumInterface::initTestCase_data(){
 
 }
 
 /**
  *  will be called after the last test function was executed.
 */
-void TestTestbed2NestedStruct3Interface::cleanupTestCase(){
+void TestTbEnumEnumInterface::cleanupTestCase(){
 
 }
 
 /**
  *  will be called before each test function is executed.
 */
-void TestTestbed2NestedStruct3Interface::init(){
+void TestTbEnumEnumInterface::init(){
 
 }
 
 /**
  *  will be called after every test function.
 */
-void TestTestbed2NestedStruct3Interface::cleanup(){
+void TestTbEnumEnumInterface::cleanup(){
 
+}
+/**
+ *  Test operation func0
+*/
+void TestTbEnumEnumInterface::func0(){
+    tb_enum::EnumInterface testTbenumenuminterface;
+    // Do implement test here
 }
 /**
  *  Test operation func1
 */
-void TestTestbed2NestedStruct3Interface::func1(){
-    testbed2::NestedStruct3Interface testTestbed2nestedstruct3interface;
+void TestTbEnumEnumInterface::func1(){
+    tb_enum::EnumInterface testTbenumenuminterface;
     // Do implement test here
 }
 /**
  *  Test operation func2
 */
-void TestTestbed2NestedStruct3Interface::func2(){
-    testbed2::NestedStruct3Interface testTestbed2nestedstruct3interface;
+void TestTbEnumEnumInterface::func2(){
+    tb_enum::EnumInterface testTbenumenuminterface;
     // Do implement test here
 }
 /**
  *  Test operation func3
 */
-void TestTestbed2NestedStruct3Interface::func3(){
-    testbed2::NestedStruct3Interface testTestbed2nestedstruct3interface;
+void TestTbEnumEnumInterface::func3(){
+    tb_enum::EnumInterface testTbenumenuminterface;
+    // Do implement test here
+}
+/**
+ *  Test proptery prop0
+*/
+void TestTbEnumEnumInterface::prop0(){
+    tb_enum::EnumInterface testTbenumenuminterface;
     // Do implement test here
 }
 /**
  *  Test proptery prop1
 */
-void TestTestbed2NestedStruct3Interface::prop1(){
-    testbed2::NestedStruct3Interface testTestbed2nestedstruct3interface;
+void TestTbEnumEnumInterface::prop1(){
+    tb_enum::EnumInterface testTbenumenuminterface;
     // Do implement test here
 }
 /**
  *  Test proptery prop2
 */
-void TestTestbed2NestedStruct3Interface::prop2(){
-    testbed2::NestedStruct3Interface testTestbed2nestedstruct3interface;
+void TestTbEnumEnumInterface::prop2(){
+    tb_enum::EnumInterface testTbenumenuminterface;
     // Do implement test here
 }
 /**
  *  Test proptery prop3
 */
-void TestTestbed2NestedStruct3Interface::prop3(){
-    testbed2::NestedStruct3Interface testTestbed2nestedstruct3interface;
+void TestTbEnumEnumInterface::prop3(){
+    tb_enum::EnumInterface testTbenumenuminterface;
     // Do implement test here
 }
 
-QTEST_MAIN(TestTestbed2NestedStruct3Interface);
+QTEST_MAIN(TestTbEnumEnumInterface);
