+++ goldenmaster/tb_same1/implementation/tests/test_sameenum2interface.cpp
@@ -15,72 +15,72 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "tst_samestruct2interface.h"
-#include "samestruct2interface.h"
+#include "test_sameenum2interface.h"
+#include "implementation/sameenum2interface.h"
 
 /**
  *  will be called before the first test function is executed.
 */
-void TestTbSame2SameStruct2Interface::initTestCase(){
+void TestTbSame1SameEnum2Interface::initTestCase(){
 
 }
 
 /**
  *  will be called to create a global test data table.
 */
-void TestTbSame2SameStruct2Interface::initTestCase_data(){
+void TestTbSame1SameEnum2Interface::initTestCase_data(){
 
 }
 
 /**
  *  will be called after the last test function was executed.
 */
-void TestTbSame2SameStruct2Interface::cleanupTestCase(){
+void TestTbSame1SameEnum2Interface::cleanupTestCase(){
 
 }
 
 /**
  *  will be called before each test function is executed.
 */
-void TestTbSame2SameStruct2Interface::init(){
+void TestTbSame1SameEnum2Interface::init(){
 
 }
 
 /**
  *  will be called after every test function.
 */
-void TestTbSame2SameStruct2Interface::cleanup(){
+void TestTbSame1SameEnum2Interface::cleanup(){
 
 }
 /**
  *  Test operation func1
 */
-void TestTbSame2SameStruct2Interface::func1(){
-    tb.same2::SameStruct2Interface testTbsame2samestruct2interface;
+void TestTbSame1SameEnum2Interface::func1(){
+    tb_same1::SameEnum2Interface testTbsame1sameenum2interface;
     // Do implement test here
 }
 /**
  *  Test operation func2
 */
-void TestTbSame2SameStruct2Interface::func2(){
-    tb.same2::SameStruct2Interface testTbsame2samestruct2interface;
+void TestTbSame1SameEnum2Interface::func2(){
+    tb_same1::SameEnum2Interface testTbsame1sameenum2interface;
     // Do implement test here
 }
 /**
  *  Test proptery prop1
 */
-void TestTbSame2SameStruct2Interface::prop1(){
-    tb.same2::SameStruct2Interface testTbsame2samestruct2interface;
+void TestTbSame1SameEnum2Interface::prop1(){
+    tb_same1::SameEnum2Interface testTbsame1sameenum2interface;
     // Do implement test here
 }
 /**
  *  Test proptery prop2
 */
-void TestTbSame2SameStruct2Interface::prop2(){
-    tb.same2::SameStruct2Interface testTbsame2samestruct2interface;
+void TestTbSame1SameEnum2Interface::prop2(){
+    tb_same1::SameEnum2Interface testTbsame1sameenum2interface;
     // Do implement test here
 }
 
-QTEST_MAIN(TestTbSame2SameStruct2Interface);
+QTEST_MAIN(TestTbSame1SameEnum2Interface);
