+++ goldenmaster/tb_same1/implementation/tests/test_sameenum2interface.h
