+++ goldenmaster/tb_same1/implementation/tests/test_samestruct2interface.cpp
@@ -15,72 +15,72 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "tst_sameenum2interface.h"
-#include "sameenum2interface.h"
+#include "test_samestruct2interface.h"
+#include "implementation/samestruct2interface.h"
 
 /**
  *  will be called before the first test function is executed.
 */
-void TestTbSame2SameEnum2Interface::initTestCase(){
+void TestTbSame1SameStruct2Interface::initTestCase(){
 
 }
 
 /**
  *  will be called to create a global test data table.
 */
-void TestTbSame2SameEnum2Interface::initTestCase_data(){
+void TestTbSame1SameStruct2Interface::initTestCase_data(){
 
 }
 
 /**
  *  will be called after the last test function was executed.
 */
-void TestTbSame2SameEnum2Interface::cleanupTestCase(){
+void TestTbSame1SameStruct2Interface::cleanupTestCase(){
 
 }
 
 /**
  *  will be called before each test function is executed.
 */
-void TestTbSame2SameEnum2Interface::init(){
+void TestTbSame1SameStruct2Interface::init(){
 
 }
 
 /**
  *  will be called after every test function.
 */
-void TestTbSame2SameEnum2Interface::cleanup(){
+void TestTbSame1SameStruct2Interface::cleanup(){
 
 }
 /**
  *  Test operation func1
 */
-void TestTbSame2SameEnum2Interface::func1(){
-    tb.same2::SameEnum2Interface testTbsame2sameenum2interface;
+void TestTbSame1SameStruct2Interface::func1(){
+    tb_same1::SameStruct2Interface testTbsame1samestruct2interface;
     // Do implement test here
 }
 /**
  *  Test operation func2
 */
-void TestTbSame2SameEnum2Interface::func2(){
-    tb.same2::SameEnum2Interface testTbsame2sameenum2interface;
+void TestTbSame1SameStruct2Interface::func2(){
+    tb_same1::SameStruct2Interface testTbsame1samestruct2interface;
     // Do implement test here
 }
 /**
  *  Test proptery prop1
 */
-void TestTbSame2SameEnum2Interface::prop1(){
-    tb.same2::SameEnum2Interface testTbsame2sameenum2interface;
+void TestTbSame1SameStruct2Interface::prop1(){
+    tb_same1::SameStruct2Interface testTbsame1samestruct2interface;
     // Do implement test here
 }
 /**
  *  Test proptery prop2
 */
-void TestTbSame2SameEnum2Interface::prop2(){
-    tb.same2::SameEnum2Interface testTbsame2sameenum2interface;
+void TestTbSame1SameStruct2Interface::prop2(){
+    tb_same1::SameStruct2Interface testTbsame1samestruct2interface;
     // Do implement test here
 }
 
-QTEST_MAIN(TestTbSame2SameEnum2Interface);
+QTEST_MAIN(TestTbSame1SameStruct2Interface);
