+++ goldenmaster/tb_same1/implementation/tests/test_samestruct2interface.h
