+++ goldenmaster/tb_same2/implementation/tests/test_sameenum2interface.cpp
@@ -15,72 +15,72 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "tst_nestedstruct2interface.h"
-#include "nestedstruct2interface.h"
+#include "test_sameenum2interface.h"
+#include "implementation/sameenum2interface.h"
 
 /**
  *  will be called before the first test function is executed.
 */
-void TestTestbed2NestedStruct2Interface::initTestCase(){
+void TestTbSame2SameEnum2Interface::initTestCase(){
 
 }
 
 /**
  *  will be called to create a global test data table.
 */
-void TestTestbed2NestedStruct2Interface::initTestCase_data(){
+void TestTbSame2SameEnum2Interface::initTestCase_data(){
 
 }
 
 /**
  *  will be called after the last test function was executed.
 */
-void TestTestbed2NestedStruct2Interface::cleanupTestCase(){
+void TestTbSame2SameEnum2Interface::cleanupTestCase(){
 
 }
 
 /**
  *  will be called before each test function is executed.
 */
-void TestTestbed2NestedStruct2Interface::init(){
+void TestTbSame2SameEnum2Interface::init(){
 
 }
 
 /**
  *  will be called after every test function.
 */
-void TestTestbed2NestedStruct2Interface::cleanup(){
+void TestTbSame2SameEnum2Interface::cleanup(){
 
 }
 /**
  *  Test operation func1
 */
-void TestTestbed2NestedStruct2Interface::func1(){
-    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
+void TestTbSame2SameEnum2Interface::func1(){
+    tb_same2::SameEnum2Interface testTbsame2sameenum2interface;
     // Do implement test here
 }
 /**
  *  Test operation func2
 */
-void TestTestbed2NestedStruct2Interface::func2(){
-    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
+void TestTbSame2SameEnum2Interface::func2(){
+    tb_same2::SameEnum2Interface testTbsame2sameenum2interface;
     // Do implement test here
 }
 /**
  *  Test proptery prop1
 */
-void TestTestbed2NestedStruct2Interface::prop1(){
-    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
+void TestTbSame2SameEnum2Interface::prop1(){
+    tb_same2::SameEnum2Interface testTbsame2sameenum2interface;
     // Do implement test here
 }
 /**
  *  Test proptery prop2
 */
-void TestTestbed2NestedStruct2Interface::prop2(){
-    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
+void TestTbSame2SameEnum2Interface::prop2(){
+    tb_same2::SameEnum2Interface testTbsame2sameenum2interface;
     // Do implement test here
 }
 
-QTEST_MAIN(TestTestbed2NestedStruct2Interface);
+QTEST_MAIN(TestTbSame2SameEnum2Interface);
