+++ goldenmaster/tb_same2/implementation/tests/test_sameenum2interface.h
