+++ goldenmaster/tb_same2/implementation/tests/test_samestruct2interface.h
