+++ goldenmaster/tb_simple/implementation/CMakeLists.txt
@@ -29,24 +29,4 @@ target_include_directories(tb_simple_impl
 target_link_libraries(tb_simple_impl PRIVATE Qt5::Core Qt5::Qml Qt5::WebSockets tb_simple::tb_simple_api)
 target_compile_definitions(tb_simple_impl PRIVATE TB_SIMPLE_IMPL_LIBRARY)
 
-# add test targets
-
-## test for SimpleInterface
-#add_executable(test_tb_simple_impl_simpleinterface
-#    tst_simpleinterface.cpp
-#)
-#
-#target_link_libraries(test_tb_simple_impl_simpleinterface PRIVATE Qt5::Test tb_simple::tb_simple_impl)
-#target_compile_definitions(test_tb_simple_impl_simpleinterface PRIVATE TB_SIMPLE_IMPL_LIBRARY)
-#add_test(test_tb_simple_impl_simpleinterface test_tb_simple_impl_simpleinterface)
-#
-
-## test for SimpleArrayInterface
-#add_executable(test_tb_simple_impl_simplearrayinterface
-#    tst_simplearrayinterface.cpp
-#)
-#
-#target_link_libraries(test_tb_simple_impl_simplearrayinterface PRIVATE Qt5::Test tb_simple::tb_simple_impl)
-#target_compile_definitions(test_tb_simple_impl_simplearrayinterface PRIVATE TB_SIMPLE_IMPL_LIBRARY)
-#add_test(test_tb_simple_impl_simplearrayinterface test_tb_simple_impl_simplearrayinterface)
-#
+add_subdirectory(tests)
