+++ goldenmaster/testbed1/implementation/CMakeLists.txt
@@ -29,24 +29,4 @@ target_include_directories(testbed1_impl
 target_link_libraries(testbed1_impl PRIVATE Qt5::Core Qt5::Qml Qt5::WebSockets testbed1::testbed1_api)
 target_compile_definitions(testbed1_impl PRIVATE TESTBED1_IMPL_LIBRARY)
 
-# add test targets
-
-## test for StructInterface
-#add_executable(test_testbed1_impl_structinterface
-#    tst_structinterface.cpp
-#)
-#
-#target_link_libraries(test_testbed1_impl_structinterface PRIVATE Qt5::Test testbed1::testbed1_impl)
-#target_compile_definitions(test_testbed1_impl_structinterface PRIVATE TESTBED1_IMPL_LIBRARY)
-#add_test(test_testbed1_impl_structinterface test_testbed1_impl_structinterface)
-#
-
-## test for StructArrayInterface
-#add_executable(test_testbed1_impl_structarrayinterface
-#    tst_structarrayinterface.cpp
-#)
-#
-#target_link_libraries(test_testbed1_impl_structarrayinterface PRIVATE Qt5::Test testbed1::testbed1_impl)
-#target_compile_definitions(test_testbed1_impl_structarrayinterface PRIVATE TESTBED1_IMPL_LIBRARY)
-#add_test(test_testbed1_impl_structarrayinterface test_testbed1_impl_structarrayinterface)
-#
+add_subdirectory(tests)
