+++ goldenmaster/testbed2/implementation/tests/test_nestedstruct2interface.cpp
@@ -15,58 +15,72 @@ GNU General Public License for more details.
 You should have received a copy of the GNU General Public License
 along with this program.  If not, see <https://www.gnu.org/licenses/>.
 */
-#include "tst_nestedstruct1interface.h"
-#include "nestedstruct1interface.h"
+#include "test_nestedstruct2interface.h"
+#include "implementation/nestedstruct2interface.h"
 
 /**
  *  will be called before the first test function is executed.
 */
-void TestTestbed2NestedStruct1Interface::initTestCase(){
+void TestTestbed2NestedStruct2Interface::initTestCase(){
 
 }
 
 /**
  *  will be called to create a global test data table.
 */
-void TestTestbed2NestedStruct1Interface::initTestCase_data(){
+void TestTestbed2NestedStruct2Interface::initTestCase_data(){
 
 }
 
 /**
  *  will be called after the last test function was executed.
 */
-void TestTestbed2NestedStruct1Interface::cleanupTestCase(){
+void TestTestbed2NestedStruct2Interface::cleanupTestCase(){
 
 }
 
 /**
  *  will be called before each test function is executed.
 */
-void TestTestbed2NestedStruct1Interface::init(){
+void TestTestbed2NestedStruct2Interface::init(){
 
 }
 
 /**
  *  will be called after every test function.
 */
-void TestTestbed2NestedStruct1Interface::cleanup(){
+void TestTestbed2NestedStruct2Interface::cleanup(){
 
 }
 /**
  *  Test operation func1
 */
-void TestTestbed2NestedStruct1Interface::func1(){
-    testbed2::NestedStruct1Interface testTestbed2nestedstruct1interface;
+void TestTestbed2NestedStruct2Interface::func1(){
+    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
+    // Do implement test here
+}
+/**
+ *  Test operation func2
+*/
+void TestTestbed2NestedStruct2Interface::func2(){
+    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
     // Do implement test here
 }
 /**
  *  Test proptery prop1
 */
-void TestTestbed2NestedStruct1Interface::prop1(){
-    testbed2::NestedStruct1Interface testTestbed2nestedstruct1interface;
+void TestTestbed2NestedStruct2Interface::prop1(){
+    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
+    // Do implement test here
+}
+/**
+ *  Test proptery prop2
+*/
+void TestTestbed2NestedStruct2Interface::prop2(){
+    testbed2::NestedStruct2Interface testTestbed2nestedstruct2interface;
     // Do implement test here
 }
 
-QTEST_MAIN(TestTestbed2NestedStruct1Interface);
+QTEST_MAIN(TestTestbed2NestedStruct2Interface);
