+++ templates/apigear/tests/olink/CMakeLists.txt
@@ -21,8 +21,6 @@ enable_testing()
 Include(FetchContent)
 
 find_package(Qt5 REQUIRED COMPONENTS Test Core Qml Network WebSockets)
-#message(STATUS "DOROTA ${CMAKE_CURRENT_LIST_DIR}")
-#SET(apigear_DIR ${CMAKE_CURRENT_LIST_DIR}/../../olink)
 find_package(apigear QUIET COMPONENTS olink_qt)
 
 FetchContent_Declare(
